+++ codigos/javascript/fav.js
@@ -5,7 +5,7 @@ export function obterFavs()
     return favsJSON ? JSON.parse(favsJSON) : []
 }
 
-export function saveFav(favoritos)
+function saveFav(favoritos)
 {
     localStorage.setItem('favoritos', JSON.stringify(favoritos));
 }
@@ -16,4 +16,22 @@ export function irFavs()
     localStorage.setItem('filtrarFavoritos', 'true');
 
     window.location.href = 'pokeall.html'
+}
+
+
+export function favoritar(numDex)
+{
+    numDex = parseInt(numDex, 10);
+    let favoritos = obterFavs();
+
+    const index = favoritos.indexOf(numDex);
+    if(index > -1)
+    {
+        favoritos.splice(index, 1);
+        console.log(`Pokemon #${numDex} removido dos favoritos`)
+    } else {
+        favoritos.push(numDex);
+        console.log(`Pokemon #${numDex} adicionado aos favoritos`)
+    }
+    saveFav(favoritos);
 }
