+++ codigos/poke.html
@@ -33,7 +33,7 @@
             <div class="linkBar">
                     <button><i class="fa-solid fa-house"></i>home</button>
                     <button onclick="window.location.href='pokeall.html'"><i class="fa-solid fa-book"></i>Pokédex</button>
-                    <button onclick="irFavs()"><i class="fa-solid fa-star"></i>Favoritos</button>
+                    <button onclick="irFavs()"><i class="fa-solid fa-star"></i>Favoritos</button><!--quando for os modulos colocar o id id="btnFavoritos" no btn-->
                     <button><i class="fa-solid fa-user"></i></button>
                 </div>
             </div>
